// packages/devtron-fe-common-lib/src/Shared/Components/EditableTextArea/types.ts
/*
 * Copyright (c) 2024. Devtron Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

interface ValidationRule<T> {
    /**
     * Value against which the input should be validated
     */
    value: T
    /**
     * Validation message when the validation fails
     */
    message: string
}

export interface EditableTextAreaProps {
    placeholder?: string
    rows: number
    updateContent: (value: string) => Promise<void>
    initialText?: string
    validations?: {
        maxLength?: ValidationRule<number>
    }
    emptyState: string
}

export interface Error {
    isValid: boolean
    message: string
}